// p14-integration/src/model/point.ts
/**
 * Universidad de La Laguna
 * Escuela Superior de Ingeniería y Tecnología
 * Grado en Ingeniería Informática
 * Programación de Aplicaciones Interactivas 2024-2025
 *
 * @author Eric Ríos Hamilton
 * @since Mar 21 2025
 * @desc Point interface representing a point in 2D space.
 */

export interface Point {
  xPosition: number;
  yPosition: number;
}